# docs/specification/data_api/openapi_data.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PlanBot Bank API Simulator",
    "description": "Raw client, holding, and product data — a stand-in for bank-internal systems.",
    "version": "0.1.0"
  },
  "paths": {
    "/api/v1/clients": {
      "get": {
        "summary": "List Clients",
        "description": "List raw client rows.\n\n``client_id`` accepts a **comma-separated** list to fetch multiple clients\nin one call.  Example: `?client_id=PB-HK-000007-5,PB-HK-000002-6`",
        "operationId": "list_clients_api_v1_clients_get",
        "parameters": [
          {
            "name": "client_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Comma-separated list of client IDs to filter by (e.g. `PB-HK-000007-5,PB-HK-000002-6`). Omit to return all clients.",
              "title": "Client Id"
            },
            "description": "Comma-separated list of client IDs to filter by (e.g. `PB-HK-000007-5,PB-HK-000002-6`). Omit to return all clients.",
            "examples": {
              "Akira Tanaka (bond holder)": {
                "value": "PB-HK-000007-5"
              },
              "Two clients (comma-separated)": {
                "value": "PB-HK-000007-5,PB-HK-000002-6"
              }
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Number of rows to skip (0-based).",
              "examples": [
                0
              ],
              "default": 0,
              "title": "Offset"
            },
            "description": "Number of rows to skip (0-based)."
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "description": "Maximum number of rows to return. Omit for all.",
              "examples": [
                10
              ],
              "title": "Limit"
            },
            "description": "Maximum number of rows to return. Omit for all."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ClientRow"
                  },
                  "title": "Response List Clients Api V1 Clients Get"
                },
                "example": [
                  {
                    "client_id": "PB-HK-000007-5",
                    "name": "Akira Tanaka",
                    "aum": 28000000.0,
                    "cash_pct": 12.0,
                    "region": "APAC",
                    "birthdate": "1980-01-01",
                    "occupation": "Real Estate Developer",
                    "risk_rating": 4,
                    "marital_status": "Single",
                    "children_info": "2 children",
                    "liquidity_need": "Low",
                    "income_stability": "Stable salaried income",
                    "investment_objective": "Long-term capital growth",
                    "qualitative_profile": "Self-made real estate developer with entrepreneurial mindset. Comfortable with illiquid and alternative investments."
                  }
                ]
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/clients/{client_id}": {
      "get": {
        "summary": "Get Client",
        "description": "Return a single raw client row.",
        "operationId": "get_client_api_v1_clients__client_id__get",
        "parameters": [
          {
            "name": "client_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Client identifier.",
              "title": "Client Id"
            },
            "description": "Client identifier.",
            "examples": {
              "Akira Tanaka (bond holder)": {
                "value": "PB-HK-000007-5"
              },
              "Sarah Chen": {
                "value": "PB-HK-000002-6"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClientRow"
                },
                "example": {
                  "client_id": "PB-HK-000007-5",
                  "name": "Akira Tanaka",
                  "aum": 28000000.0,
                  "cash_pct": 12.0,
                  "region": "APAC",
                  "birthdate": "1980-01-01",
                  "occupation": "Real Estate Developer",
                  "risk_rating": 4,
                  "marital_status": "Single",
                  "children_info": "2 children",
                  "liquidity_need": "Low",
                  "income_stability": "Stable salaried income",
                  "investment_objective": "Long-term capital growth",
                  "qualitative_profile": "Self-made real estate developer with entrepreneurial mindset. Comfortable with illiquid and alternative investments."
                }
              }
            }
          },
          "404": {
            "description": "Client not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/holdings": {
      "get": {
        "summary": "List Holdings",
        "description": "List raw holding rows.\n\n``client_id`` accepts a **comma-separated** list to fetch holdings for\nmultiple clients in one call.",
        "operationId": "list_holdings_api_v1_holdings_get",
        "parameters": [
          {
            "name": "client_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Comma-separated list of client IDs to filter by (e.g. `PB-HK-000007-5`). Omit to return all holdings.",
              "title": "Client Id"
            },
            "description": "Comma-separated list of client IDs to filter by (e.g. `PB-HK-000007-5`). Omit to return all holdings.",
            "examples": {
              "Akira Tanaka (bond holder)": {
                "value": "PB-HK-000007-5"
              },
              "Two clients (comma-separated)": {
                "value": "PB-HK-000007-5,PB-HK-000002-6"
              }
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Number of rows to skip (0-based).",
              "examples": [
                0
              ],
              "default": 0,
              "title": "Offset"
            },
            "description": "Number of rows to skip (0-based)."
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "description": "Maximum number of rows to return. Omit for all.",
              "examples": [
                20
              ],
              "title": "Limit"
            },
            "description": "Maximum number of rows to return. Omit for all."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HoldingRow"
                  },
                  "title": "Response List Holdings Api V1 Holdings Get"
                },
                "example": [
                  {
                    "client_id": "PB-HK-000007-5",
                    "holding_idx": 0,
                    "holding_id": "ph-6-us1mt-rr-0",
                    "product_id": "PROD053",
                    "instrument_name": "US 1-Month Treasury Bill Rate",
                    "symbol": "US1MT=RR",
                    "asset_class": "Cash",
                    "region": "North America",
                    "currency": "USD",
                    "quantity": 153494.7465,
                    "book_cost": 2352941.1765,
                    "market_value": 3360000.0,
                    "unrealized_pl": 1007058.8235,
                    "unrealized_pl_pct": 42.8,
                    "yield_pct": 15.7,
                    "risk_bucket": "Low",
                    "liquidity": "T+2"
                  }
                ]
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/products": {
      "get": {
        "summary": "List Products",
        "description": "List raw product rows.\n\n``product_id`` accepts a **comma-separated** list to fetch multiple products\nin one call.  Example: `?product_id=PROD053,PROD054`",
        "operationId": "list_products_api_v1_products_get",
        "parameters": [
          {
            "name": "product_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Comma-separated list of product IDs to filter by (e.g. `PROD053,ETF-HYG`). Omit to return the full product catalog.",
              "title": "Product Id"
            },
            "description": "Comma-separated list of product IDs to filter by (e.g. `PROD053,ETF-HYG`). Omit to return the full product catalog.",
            "examples": {
              "US Treasury 4.375% 31Aug26": {
                "value": "PROD053"
              },
              "Two products (comma-separated)": {
                "value": "PROD053,ETF-HYG"
              }
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Number of rows to skip (0-based).",
              "examples": [
                0
              ],
              "default": 0,
              "title": "Offset"
            },
            "description": "Number of rows to skip (0-based)."
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "description": "Maximum number of rows to return. Omit for all.",
              "examples": [
                10
              ],
              "title": "Limit"
            },
            "description": "Maximum number of rows to return. Omit for all."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductRow"
                  },
                  "title": "Response List Products Api V1 Products Get"
                },
                "example": [
                  {
                    "product_id": "PROD053",
                    "name": "US Treasury 4.375% 31Aug26",
                    "ticker": "XHLF",
                    "trading_currency": "USD",
                    "risk_rating": 1,
                    "expected_return": 3.7,
                    "region": "US",
                    "sector": "Government",
                    "remarks": "Bridge row for holdings product_id=PROD053. Group: holdings_treasury_bonds.",
                    "product_type": "bond",
                    "vehicle": "Direct",
                    "type_specific": {
                      "issuer_name": "U.S. Treasury",
                      "issuer_sector": "government",
                      "coupon_type": "fixed",
                      "coupon_rate": 0.04375,
                      "coupon_frequency": "semi-annual",
                      "credit_rating": "AA+",
                      "maturity": "2026-08-31",
                      "seniority": "senior",
                      "callable": false
                    },
                    "performance_history": {
                      "1y": {
                        "return": 3.65,
                        "cagr": 3.72,
                        "max_drawdown": 0.0,
                        "volatility": 0.2
                      },
                      "3y": {
                        "return": 14.22,
                        "cagr": 4.53,
                        "max_drawdown": -0.03,
                        "volatility": 0.28
                      }
                    },
                    "investment_note": "Individual bonds allow precise maturity-matching for liability-driven investing."
                  }
                ]
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/products/{product_id}": {
      "get": {
        "summary": "Get Product",
        "description": "Return a single raw product row.",
        "operationId": "get_product_api_v1_products__product_id__get",
        "parameters": [
          {
            "name": "product_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Product identifier.",
              "title": "Product Id"
            },
            "description": "Product identifier.",
            "examples": {
              "US Treasury 4.375% 31Aug26": {
                "value": "PROD053"
              },
              "High Yield Corporate Bond ETF": {
                "value": "ETF-HYG"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductRow"
                },
                "example": {
                  "product_id": "PROD053",
                  "name": "US Treasury 4.375% 31Aug26",
                  "ticker": "XHLF",
                  "trading_currency": "USD",
                  "risk_rating": 1,
                  "expected_return": 3.7,
                  "region": "US",
                  "sector": "Government",
                  "remarks": "Bridge row for holdings product_id=PROD053. Group: holdings_treasury_bonds.",
                  "product_type": "bond",
                  "vehicle": "Direct",
                  "type_specific": {
                    "issuer_name": "U.S. Treasury",
                    "issuer_sector": "government",
                    "coupon_type": "fixed",
                    "coupon_rate": 0.04375,
                    "coupon_frequency": "semi-annual",
                    "credit_rating": "AA+",
                    "maturity": "2026-08-31",
                    "seniority": "senior",
                    "callable": false
                  },
                  "performance_history": {
                    "1y": {
                      "return": 3.65,
                      "cagr": 3.72,
                      "max_drawdown": 0.0,
                      "volatility": 0.2
                    },
                    "3y": {
                      "return": 14.22,
                      "cagr": 4.53,
                      "max_drawdown": -0.03,
                      "volatility": 0.28
                    }
                  },
                  "investment_note": "Individual bonds allow precise maturity-matching for liability-driven investing."
                }
              }
            }
          },
          "404": {
            "description": "Product not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ErrorDetail": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail",
            "description": "Human-readable error message."
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "ErrorDetail",
        "description": "Standard error body returned on 404."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}